server/pyproject.toml
[tool.poetry]
name = "dj-grpc-benchmark"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Django = "2.2.17"
grpcio-tools = "^1.43.0"
djangorestframework = "^3.13.1"
djangogrpcframework = "^0.2.1"

[tool.poetry.dev-dependencies]
psycopg2-binary = "~=2.9.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
